decompose ChatPage into components and add date management
